🎨 #128 nav 목표 추가 버튼 위치 조정
- 모바일과 태블릿/데스크탑에서 목표 추가 버튼의 위치를 조절했습니다.
  - 하나의 목표 추가 컴포넌트를 이용하도록 수정했습니다.
